+++ UVtools.Core/FileFormats/PWSFile.cs
@@ -803,7 +803,11 @@ namespace UVtools.Core.FileFormats
         public override uint ResolutionY => HeaderSettings.ResolutionY;
         public override byte AntiAliasing => (byte) HeaderSettings.AntiAliasing;
 
-        public override float LayerHeight => HeaderSettings.LayerHeight;
+        public override float LayerHeight
+        {
+            get => HeaderSettings.LayerHeight;
+            set => HeaderSettings.LayerHeight = value;
+        }
 
         public override uint LayerCount
         {