+++ UVtools.Core/Layer/LayerManager.cs
@@ -1437,5 +1437,64 @@ namespace UVtools.Core
             BoundingRectangle = Rectangle.Empty;
             SlicerFile.RequireFullEncode = true;
         }
+
+        public void ReHeight(OperationLayerReHeight operation, OperationProgress progress = null)
+        {
+            if (ReferenceEquals(progress, null)) progress = new OperationProgress();
+            progress.Reset("Re-Height", operation.LayerCount);
+
+            var oldLayers = Layers;
+
+            Layers = new Layer[operation.LayerCount];
+
+            uint newLayerIndex = 0;
+            for (uint layerIndex = 0; layerIndex < oldLayers.Length; layerIndex++)
+            {
+                var oldLayer = oldLayers[layerIndex];
+                if (operation.IsDivision)
+                {
+                    for (byte i = 0; i < operation.Modifier; i++)
+                    {
+                        Layers[newLayerIndex] =
+                            new Layer(newLayerIndex, oldLayer.CompressedBytes, null, this)
+                            {
+                                PositionZ = (float) (operation.LayerHeight * (newLayerIndex + 1)),
+                                ExposureTime = oldLayer.ExposureTime,
+                                BoundingRectangle = oldLayer.BoundingRectangle,
+                            };
+                        newLayerIndex++;
+                        progress++;
+                    }
+                }
+                else
+                {
+                    using (var mat = oldLayers[layerIndex++].LayerMat)
+                    {
+                        for (byte i = 1; i < operation.Modifier; i++)
+                        {
+                            using (var nextMat = oldLayers[layerIndex++].LayerMat)
+                            {
+                                CvInvoke.Add(mat, nextMat, mat);
+                            }
+                        }
+
+                        Layers[newLayerIndex] = new Layer(newLayerIndex, mat, null, this)
+                        {
+                            PositionZ = (float)(operation.LayerHeight * (newLayerIndex + 1)),
+                            ExposureTime = oldLayer.ExposureTime
+                        };
+                        newLayerIndex++;
+                        layerIndex--;
+                        progress++;
+                    }
+                }
+            }
+
+
+            SlicerFile.LayerHeight = (float) operation.LayerHeight;
+            SlicerFile.LayerCount = Count;
+            BoundingRectangle = Rectangle.Empty;
+            SlicerFile.RequireFullEncode = true;
+        }
     }
 }