+++ UVtools.GUI/Forms/FrmInputBox.cs
@@ -7,6 +7,7 @@
  */
 
 using System;
+using System.Diagnostics;
 using System.Globalization;
 using System.Windows.Forms;
 using UVtools.Core;